+++ src/util.h
@@ -122,12 +122,12 @@ char *reverse(const char *s, char**buffer, size_t*n);
 /* - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
    This function is a replacement for POSIX.1-2008 getdelim()
 */
-size_t readdelim(char **lineptr, size_t  *n, int delim, FILE *stream);
+int readdelim(char **lineptr, size_t *n, int delim, FILE *stream);
 
 /* - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
    This function is a replacement for POSIX.1-2008 getline()
 */
-size_t readline(char **lineptr, size_t  *n, FILE *stream);
+int readline(char **lineptr, size_t *n, FILE *stream);
 
 /* - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
    successively returns new lines from file, NOT reentrant
